[handles] Use is_assignable instead of is_base_of
std::is_base_of<A, B> has undefined behaviour if B is not a complete
type. Hence, avoid it and use is_assignable<A*&, B*> instead.
This captures exactly the requirement that
  Handle<A> x = produce<Handle<B>>();
is valid exactly if
  A* x = produce<B*>();
is valid.
Also, change some static asserts to enable_ifs in order to allow
overloading a function by different Handle types, which would cause
disambiguity otherwise.

R=tebbi@chromium.org

Change-Id: I66c66db1567f647a44e047fff3c36385898887a8