[ic] Convert undefined key to undefined_value.
In the special case of KeyedLoadIC, where the key is always undefined,
LowerCheckEqualsInternalizedString fails and triggers deoptimization
loops. It's because {exp}, which is undefined_string, does not match
{val}, which is undefined_value.

This patch changes the convertion target from undefined_string to
undefined_value for undefined key in KeyedLoadIC. It makes undefined
key be able to pass LowerCheckEqualsInternalizedString and avoids the
deoptimization loops.

Bug:
Change-Id: Ie15358e705490fb2e4907f4d37a3ef7921f74844