Reland [platform][cleanup] Fix --enable-tracing
The reason for the revert was a data race between the execution of the
Exit() function and other threads which also access the platform, e.g.
the IsolateThreads and the Sampling thread which is attached to the
isolate. With this CL we wait for these threads before we shut down the
platform.

I think this behavior is better than the original behavior where
--enable-tracing was non-functional when d8 finished with a call to Exit()
or with an uncaught exception. For example, the Unity benchmark ends with
a call to Exit() and was therefore difficult to analyze before this
change. However, you can also argue that Exit() should just exit the
application immediately, and waiting for threads is not part of the deal.

Original change's description:
> [platform][cleanup] Fix --enable-tracing
>
> The flag --enable-tracing can be used to measure how much time is spent
> in a scope. In d8 this flag did not work properly for the following
> reasons:
>
> * The tracing file is not written when the JavaScript code calls quit().
>
> * The tracing file is not written when the JavaScript code throws an
> uncaught exception, except if the --throws flag is passed to d8.
>
> The reason for these two issues is that both call Shell::Exit(), which
> end d8 immediately without calling any destructors. In addition I moved
> in a recent CL the destruction of the platform after the destruction of
> the file handle for the tracing file. Thereby the tracing file did not
> get filled with content even destructors were executed.
>
> With this CL I also call the destructors of the platform and of the file
> handle in the Shell::Exit() function. For this I make Shell::Exit()
> platform independent and let it call the platform dependent
> Shell::OSExit() at the end.
>
> R=rmcilroy@chromium.org
>
> Change-Id: I14425b6a5c25b757211bc8b9959a9cc8bfa0602c
> Reviewed-on: https://chromium-review.googlesource.com/789038
> Reviewed-by: Ross McIlroy <rmcilroy@chromium.org>
> Commit-Queue: Andreas Haas <ahaas@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#49659}

Change-Id: Icb0c47af7bf1536cc835a5bfdcbe6924643965ca