Update patch set 9
Patch Set 9:

Okay, the mjsunit suite is passing at this point (other suites mostly untested with the exception of test/inspector, but more test updates are likely needed)

Currently, there's a lot of stuff here that doesn't need to be here. I'm quite sure the changes to bytecode-generator.cc can be reduced a good amount, as well as the 3 functions implemented in both the preparser and full parser for bindings variable declarations and catch parameter bindings.

Once it's properly cleaned up, this should be a net reduction in LoC, but at the moment it's a bit of an increase :)

Todo:
  - Simplify abstractions for adding catch handlers in BytecodeGenerator
  - Simplify abstractions for the context chain in BytecodeGenerator
  - Simplify and add more abstractions for iterator uses
  - Delete more dead parser functions
  - Cleanup new parser functions relating to scope declarations
  - Anything else people think is a good idea
  - Delete ASTTraversalVisitor and ASTExpressionRewriter if possible?
  - Do better job for prettyprinting
  - Remove whatever that change to src/base/bits.h is
  - Make sure we aren't doing more work during preparsing than before

There are definitely some good ideas here, but the execution does need some improvement. I'll spend a bit more time trying to get it as clean as possible, then rebase on top of ToT and try to pull out the simplest parts into separate patches.

Patch-set: 9
Reviewer: Gerrit User 1162379 <1162379@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false