Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Ali, disabling the test would be ok if only the allocation profiler depended on allocation profiler. Unfortunately, scavenger and incremental marker rely on allocation observers.
>
> So this bug will probably cause crashes in production for Chrome users not using allocation profiler. This is also going to generate unnecessary work for stability sheriffs investigating memory dumps.
> 
> It is painful, but we should revert the CL and reland with the fix. I can help with debugging this issue on Thursday (currently there is compiler summit in Munich).

Note that I have already debugged the problem and the CL already contains the fix - albeit commented out. Instead of revert I propose that we enable code. I believe it is correct this time around. Failures have pinpointed the precise bug and all the failures have the exact same symptoms. I am sorry it has been painful getting here. Tell me what additional tests I can run.

There was an obvious bug. You fixed it partly in the CL: https://chromium-review.googlesource.com/c/v8/v8/+/738204, but it was missing in one of the places.

> I can help with debugging this issue on Thursday

This does not need any more debugging.

I am sorry but this approach devalues my time. If we go at this pace, I am never going to be able to make progress on this. I already have to wait 24-48 hours round trip in changes - or wake up in the middle of the night to make sure I can make progress.

I have adjusted the CL to enable the assertions that I believe to be correct now.

Patch-set: 2