Update patch set 5
Patch Set 5:

> To help reduce review latency I'll check my incoming reviews at 20:00 MUC time (11:00 SFO) each workday. Please send CLs before that time to get quick reviews.
> 

Thank you, that would help make collaboration a lot more smoother. However, note that I am not specifically asking for this. I was concerned about delaying things till Thursday which would have stalled progress for me for pretty much the whole week.

> I don't know how to help with the flaky failures that show up days later downstream. 

This is a problem. For the last round it took 4 days for the failures to show up. If there is a chance that the change might be reverted, it is difficult to start work on a follow-up change because the reverts may end up cascading.

> I think we need more tests on V8 side for allocation observers.

Actually we need more test for edge cases in allocation in addition to allocation observer tests. It seems that allocation patterns at page boundaries are not well exercised.

Patch-set: 5