[wasm] Test decoder tracing output
This will be refactored in a follow-up CL https://crrev.com/c/813972), so
add a message test to make sure that we don't break something.
This will also allow us to observe the change in output in that follow-up
CL.

Drive-by: Fix testcfg.py to rstrip the actual lines, because also the
expected lines were rstriped, so they don't match if there are trailing
whitespaces (like in our new test).

R=titzer@chromium.org, machenbach@chromium.org

Change-Id: I6bab282cd10e0be74d28480d3cfb7aadecd92ae7