[interpreter] return to old iteration protocol for array spreads
This reverts to the old iteration protocol for array literal "spreads",
and loads the ".next" method every time it's called, rather than only
once.

This fixes a regression in the SixSpeed/spread_literal/es6, but
introduces behaviour which is no longer spec compliant, and is thus not
a valid solution.

A better solution would be to figure out why the current bytecode in ToT
produces such a weird graph in TurboFan.

R=rmcilroy@chromium.org, bmeurer@chromium.org, leszeks@chromium.org

BUG=chromium:795632

Change-Id: I18af7aa9ab46afe8045abdbec4e292e13e921023