Update patch set 2
Patch Set 2:

Hi, so as I mentioned on the bug, this change gets rid of the regression (or most of it -- I did see some minor improvements from other things like using the %_IsJSReceiver intrinsic rather than the JumpIfJSReceiver bytecode, but this covers the steep dive in the SixSpeed test)

As I also mentioned on the bug, a fix like this is not the correct way to solve this. I would appreciate some help or an introduction to diving into the turbofan passes which seem to cause the big problems: As noted, the big differences begin in the loop peeling phase, and culminate in very large differences in the ControlFlowOptimization phase.

Patch-set: 2
Reviewer: Gerrit User 1162379 <1162379@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false