[arm|arm64][turbofan] Refactor AssembleMove and AssembleSwap
The way the code generator's AssembleMove and AssembleSwap methods are written
makes it easy to forget which sort of move is being implemented when looking at
a sequence of instructions. This patch is an attempt to address this by
rewriting those methods using switch/case instead of a string of if/else.

To do this, introduce new utility functions to detect what type of move to
perform given a pair of InstructionOperands.

Bug:
Change-Id: I32b146c86409e595b7b59a66bf43220899024fdd