[test] Add basic test-runner system tests
This tests some high-level scenarios of the V8 test runner, with
executable, test-suite extension and build configs mocked out.

The runners are slightly modified to be easier testable. Args
are passed from the tests now and the V8 root folder can be faked
by the tests.

We support coverage if python coverage 4.0 is installed. Otherwise
we run without it.

Bug: v8:6917
Change-Id: Ib149fd88027cbdc3382bcaea2d82020582f79d2d