[cctest] Skip test-run-wasm-simd if SIMD is not supported
For mips, if 'mips_arch_variant=="r6"' and if 'mips_use_msa' flag is set
to 'true', then these tests won't be skipped for mips.
ARM processors are assumed to support SIMD.

Also, if 'mips_use_msa' build flag is 'true', it forces 'MIPS_SIMD' bit
in CpuFeatures to be set.

Bug:
Change-Id: Ib3eb7e71bf5674c2f32c6e0a54a14604da831eb4