Update patch set 2
Patch Set 2:

Hi there,

This CL gives us decent code size improvements! As it is the first time dealing with higher level bits of turbofan, I'm not 100% confident about the approach, hopefully it makes sense.

For reference, here are a couple of approaches I tried which turned to be not so great but led to this patch.

- Teach the selector to match NumberConstants if they can be represented as Smis.

This was the most obvious way to do it, and there is precedent in the x64 selector where we match NumberConstant 0 as an immediate operand. But, for it to work generally for all binops, we need to teach Int{32,64}Matcher to recognize compatible NumberConstants. This means we're changing the behaviour of those matchers in non-obvious ways, and it potentially affects all their users. This doesn't seem like a good idea.

And in general, it feels much better to just let the selector not have to worry about Smis. So I looked into lowering NumberConstants to Int{32,64}Constants when possible.

- Lower NumberConstant only when the output is used as a Smi.

A first attempt was to lower NumberConstant when used by a binop deemed to be a TaggedSigned/Smi operation, so inside the representation changer. This made sense because this is also the place where a SpeculativeNumberLessThan gets lowered to a Int{32,64}LessThan, for example.

This seemed perfect, but it turned out we were missing optimisations since lowering 
NumberConstant to a Int{32,64}Constant is valid everytime the constant can be represented as a Smi, not just according to its uses.

So that's how this CL came to be, let's lower NumberConstant everytime it can be a Smi. What do you think?

Finally, there is one little caveat for debugging: we now end up with Int64Constant nodes with big decimal numbers in them. We have to mentally untag them to know what number they actually represent. Maybe we could print them as hexadecimal to help a little.

Patch-set: 2
Reviewer: Gerrit User 1183894 <1183894@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1118629 <1118629@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1183769 <1183769@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0
Work-in-progress: false