Update patch set 2
Patch Set 2:

Just realized that this change is not useful. Will abandon. The reason is that the code-serializer test step in infra will add the --cache=code flag to all our variants, which is different from just adding one --code-cache variant.

Maybe this is still useful in the future with a feature that cross-combines variants. Will reopen this once we have it.

Patch-set: 2
Reviewer: Gerrit User 1115956 <1115956@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>