Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> OK, thanks for trying this. I am still concerned about the increase in complexity both in the public interface, as well as the implementation. Is there a particular reason why you are using templates here? I am not opposed to templates in general, but here we add lots of them, calling each other, and it seems that they rather increase complexity instead of reducing it.
> 
> We are already passing the NumericImplementation as a parameter, why not also pass the source type and target type as parameter? Instead of the template specializations, you would then just have a (non-exposed) helper method with a switch over the type. Instead of several such methods, you could even have just one, returning a struct with all the information needed (min, max, opcodes - you could even include a function for generating the zero constant).
> I imagine that resulting in much cleaner code.

We should probably come up with some guidelines for usage of templates because we tend to have many discussions in this area. We've had to judge CLs one by one and it'd be good to have some rules of thumb for guidance.

Before we have that, I was asked to take a look at this CL. I agree with Clemens in this case that templates don't add much to wasm-compiler.* here. The passed values could as well be regular arguments using a struct. We're inside a compiler, so information about (WASM) types is generally encoded as C++ data rather than C++ types. The templates in the interpreter look fine, since they are parameterized over C++ types, because they manipulate C++ values that correspond to WASM values.

Patch-set: 1
CC: Gerrit User 1119499 <1119499@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>