Update patch set 1
Patch Set 1: Code-Review+1

The fix lgtm.

As for a PreParserScopeAnalysis test, I have to admit that I don't understand that test fixture well enough to craft something of my own.

But I'll point out that the example you gave doesn't involve a named function expression whose name is "arguments"; instead, it's a function declaration whose name is arguments, which simply results in a variable named "arguments" in the containing function. The function name declaration code this patch touches only triggers for expressions, which require that "function foo..." be in an expression position, that is, not at the beginning of a statement. So that might help you craft a "better" test (the test you have at least fails without this patch, so it's "good enough" to me).

Patch-set: 1
Reviewer: Gerrit User 1115851 <1115851@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1