Update patch set 1
Patch Set 1: Code-Review+1

LGTM. Hmm, I guess the test will be fragile because you are dependent on:

a) the closure must have been inlined (inlining policy)
b) the closure must have been marked as not-escaping and was not created (escape analysis)
c) the deoptimizer must function in a particular way in how/when it creates the closure (deoptimization implementation).

With all that, I can imagine that something varies across platforms or test configurations. What's more, achieving perfect alignment here is fragile in time. Eventually policy will change.

We have a similar problem in testing whether or not a particular optimization happened in TurboFan, and to avoid creating fragile tests we simply make sure that we don't crash, rather than invasively verifying that an optimization took place.

I'd relax the test to accept that this special text string may or may not appear. The important thing is that there is no assert or crash.

Patch-set: 1
Reviewer: Gerrit User 1118564 <1118564@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1